fix(api): add UserChatEvent `#isCancellable` and `#isModifiable`
Due to changes in Minecraft 1.19.1+, some platforms no longer support
cancelling or modifying the chat event.

Add `UserChatEvent#isCancellable()` and `UserChatEvent#isModifiable()`
to allow developers to determine whether these operations are supported
on the current platform.

If `UserChatEvent#isCancellable()` returns false, but the event is
cancelled, the event will not be cancelled on the platform.

If `UserChatEvent#isModifiable()` returns false, but the chat message is
modified, the event's chat message will not be changed on the platform.