build!: cleanup adventure serializer dependence (#338)
Internally, Chameleon uses the legacy and GSON Adventure component
serializers. However, as they are not a part of the API nor are needed
by end-users, they should not be exposed as `api` dependencies.

Additionally, the legacy serializer is only required by the Bukkit and
Nukkit platform implementations, and should not be a dependency of the
API module.