build: switch to Goooler/shadow fork
The Gradle shadow plugin by johnrengelman is currently unmaintained, and
does not currently support the latest versions of Gradle.

There have been discussions regarding the future of the plugin, however
nothing has been acted upon currently: johnrengelman/shadow#908

This fork appears to be the most supported and best maintained fork, and
has been suggested as the replacement for the original shadow plugin.

Fixes #414
Related to #410